// src/gql/graphql.ts
export type Order = Node & {
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<Order>;
  email?: Maybe<Scalars['String']['output']>;
  /** List of Order versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  orderItems: Array<OrderItem>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  /** System stage field */
  stage: Stage;
  stripeCheckoutId?: Maybe<Scalars['String']['output']>;
  total: Scalars['Int']['output'];
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type Product = Node & {
  averageRating: Scalars['Int']['output'];
  categories: Array<Category>;
  collections: Array<Collection>;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description: Scalars['String']['output'];
  /** Get the document in other stages */
  documentInStages: Array<Product>;
  /** List of Product versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  images: Array<Asset>;
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Product>;
  name: Scalars['String']['output'];
  orderItems: Array<OrderItem>;
  price: Scalars['Int']['output'];
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  reviews: Array<Review>;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  variants: Array<ProductVariants>;
};
// ...
export type CartAddProductMutationVariables = Exact<{
  orderId: Scalars['ID']['input'];
  total: Scalars['Int']['input'];
  productId: Scalars['ID']['input'];
}>;


export type CartAddProductMutation = { createOrderItem?: { id: string } | null };

export type CartCreateMutationVariables = Exact<{ [key: string]: never; }>;
// ...
export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
export const ReviewsItemFragmentDoc = new TypedDocumentString(`
    fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}
    `, {"fragmentName":"ReviewsItem"}) as unknown as TypedDocumentString<ReviewsItemFragment, unknown>;
export const ProductItemFragmentDoc = new TypedDocumentString(`
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
    fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`, {"fragmentName":"ProductItem"}) as unknown as TypedDocumentString<ProductItemFragment, unknown>;
export const CartFragmentDoc = new TypedDocumentString(`
    fragment Cart on Order {
  id
  orderItems {
    id
    quantity
    total
    product {
      ...ProductItem
    }
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`, {"fragmentName":"Cart"}) as unknown as TypedDocumentString<CartFragment, unknown>;
export const CartAddProductDocument = new TypedDocumentString(`
    mutation CartAddProduct($orderId: ID!, $total: Int!, $productId: ID!) {
  createOrderItem(
    data: {quantity: 1, total: $total, product: {connect: {id: $productId}}, order: {connect: {id: $orderId}}}
  ) {
    id
  }
}
    `) as unknown as TypedDocumentString<CartAddProductMutation, CartAddProductMutationVariables>;
export const CartCreateDocument = new TypedDocumentString(`
    mutation CartCreate {
  createOrder(data: {total: 0}) {
    ...Cart
  }
}
    fragment Cart on Order {
  id
  orderItems {
    id
    quantity
    total
    product {
      ...ProductItem
    }
  }
}
fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<CartCreateMutation, CartCreateMutationVariables>;
export const CartGetByIdDocument = new TypedDocumentString(`
    query CartGetById($id: ID!) {
  order(where: {id: $id}, stage: DRAFT) {
    ...Cart
  }
}
    fragment Cart on Order {
  id
  orderItems {
    id
    quantity
    total
    product {
      ...ProductItem
    }
  }
}
fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<CartGetByIdQuery, CartGetByIdQueryVariables>;
export const CartRemoveProductDocument = new TypedDocumentString(`
    mutation CartRemoveProduct($itemId: ID!) {
  deleteOrderItem(where: {id: $itemId}) {
    id
  }
}
    `) as unknown as TypedDocumentString<CartRemoveProductMutation, CartRemoveProductMutationVariables>;
export const CartUpdateItemQuantityDocument = new TypedDocumentString(`
    mutation CartUpdateItemQuantity($itemId: ID!, $quantity: Int!) {
  updateOrderItem(where: {id: $itemId}, data: {quantity: $quantity}) {
    id
  }
}
    `) as unknown as TypedDocumentString<CartUpdateItemQuantityMutation, CartUpdateItemQuantityMutationVariables>;
export const CartUpsertProductDocument = new TypedDocumentString(`
    mutation CartUpsertProduct($cartId: ID!, $productId: ID!, $orderItemId: ID, $total: Int!, $quantity: Int!) {
  upsertOrderItem(
    upsert: {create: {quantity: 1, total: $total, order: {connect: {id: $cartId}}, product: {connect: {id: $productId}}}, update: {quantity: $quantity, total: $total}}
    where: {id: $orderItemId}
  ) {
    id
  }
}
    `) as unknown as TypedDocumentString<CartUpsertProductMutation, CartUpsertProductMutationVariables>;
export const CategoriesGetListDocument = new TypedDocumentString(`
    query CategoriesGetList {
  categories {
    id
    name
    slug
  }
}
    `) as unknown as TypedDocumentString<CategoriesGetListQuery, CategoriesGetListQueryVariables>;
export const ProductGetByIdDocument = new TypedDocumentString(`
    query ProductGetById($id: ID!) {
  product(where: {id: $id}) {
    ...ProductItem
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductGetByIdQuery, ProductGetByIdQueryVariables>;
export const ProductPublishDocument = new TypedDocumentString(`
    mutation ProductPublish($id: ID!) {
  publishProduct(where: {id: $id}, to: PUBLISHED) {
    ...ProductItem
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductPublishMutation, ProductPublishMutationVariables>;
export const ProductsGetByCategorySlugDocument = new TypedDocumentString(`
    query ProductsGetByCategorySlug($slug: String!, $first: Int, $skip: Int) {
  categories(where: {slug: $slug}) {
    name
    description
    products(first: $first, skip: $skip) {
      ...ProductItem
    }
  }
  productsConnection(where: {categories_some: {slug: $slug}}) {
    pageInfo {
      pageSize
    }
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductsGetByCategorySlugQuery, ProductsGetByCategorySlugQueryVariables>;
export const ProductsGetByCollectionSlugDocument = new TypedDocumentString(`
    query ProductsGetByCollectionSlug($slug: String!) {
  collections(where: {slug: $slug}) {
    name
    description
    products {
      ...ProductItem
    }
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductsGetByCollectionSlugQuery, ProductsGetByCollectionSlugQueryVariables>;
export const ProductsGetListDocument = new TypedDocumentString(`
    query ProductsGetList($first: Int, $skip: Int, $query: String, $orderBy: ProductOrderByInput) {
  products(
    where: {name_contains: $query}
    first: $first
    skip: $skip
    orderBy: $orderBy
  ) {
    ...ProductItem
  }
  productsConnection(where: {name_contains: $query}) {
    pageInfo {
      pageSize
    }
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductsGetListQuery, ProductsGetListQueryVariables>;
export const ReviewCreateDocument = new TypedDocumentString(`
    mutation ReviewCreate($id: ID!, $headline: String!, $name: String!, $email: String!, $content: String!, $rating: Int!) {
  createReview(
    data: {headline: $headline, name: $name, email: $email, content: $content, rating: $rating, product: {connect: {id: $id}}}
  ) {
    ...ReviewsItem
  }
}
    fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ReviewCreateMutation, ReviewCreateMutationVariables>;
export const ReviewPublishDocument = new TypedDocumentString(`
    mutation ReviewPublish($id: ID!) {
  publishReview(where: {id: $id}, to: PUBLISHED) {
    ...ReviewsItem
  }
}
    fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ReviewPublishMutation, ReviewPublishMutationVariables>;
export const ReviewUpdateAverageRatingDocument = new TypedDocumentString(`
    mutation ReviewUpdateAverageRating($id: ID!, $averageRating: Int) {
  updateProduct(where: {id: $id}, data: {averageRating: $averageRating}) {
    id
  }
}
    `) as unknown as TypedDocumentString<ReviewUpdateAverageRatingMutation, ReviewUpdateAverageRatingMutationVariables>;